Extract JavaCompile hacks into its own class
